+++ config/runtime/endpoint.go
@@ -76,6 +76,9 @@ func newEndpointOptions(confCtx *hcl.EvalContext, endpointConf *config.Endpoint,
 		errTpl = serverOptions.ServerErrTpl
 	}
 
+	// blockBodies contains inner endpoint block remain bodies to determine req/res buffer options.
+	var blockBodies []hcl.Body
+
 	var response *producer.Response
 	// var redirect producer.Redirect // TODO: configure redirect block
 	proxies := make(producer.Proxies, 0)
@@ -85,10 +88,11 @@ func newEndpointOptions(confCtx *hcl.EvalContext, endpointConf *config.Endpoint,
 		response = &producer.Response{
 			Context: endpointConf.Response.Remain,
 		}
+		blockBodies = append(blockBodies, response.Context)
 	}
 
 	for _, proxyConf := range endpointConf.Proxies {
-		backend, berr := newBackend(confCtx, proxyConf.Backend, log, proxyEnv, memStore)
+		backend, innerBody, berr := newBackend(confCtx, proxyConf.Backend, log, proxyEnv, memStore)
 		if berr != nil {
 			return nil, berr
 		}
@@ -98,10 +102,11 @@ func newEndpointOptions(confCtx *hcl.EvalContext, endpointConf *config.Endpoint,
 			RoundTrip: proxyHandler,
 		}
 		proxies = append(proxies, p)
+		blockBodies = append(blockBodies, proxyConf.Backend, innerBody, proxyConf.HCLBody())
 	}
 
 	for _, requestConf := range endpointConf.Requests {
-		backend, berr := newBackend(confCtx, requestConf.Backend, log, proxyEnv, memStore)
+		backend, innerBody, berr := newBackend(confCtx, requestConf.Backend, log, proxyEnv, memStore)
 		if berr != nil {
 			return nil, berr
 		}
@@ -111,6 +116,7 @@ func newEndpointOptions(confCtx *hcl.EvalContext, endpointConf *config.Endpoint,
 			Context: requestConf.Remain,
 			Name:    requestConf.Name,
 		})
+		blockBodies = append(blockBodies, requestConf.Backend, innerBody, requestConf.HCLBody())
 	}
 
 	backendConf := *DefaultBackendConf
@@ -138,8 +144,7 @@ func newEndpointOptions(confCtx *hcl.EvalContext, endpointConf *config.Endpoint,
 		}}
 	}
 
-	// TODO: determine request/backend_responses.*.body access in this context (all including backend) or for now:
-	bufferOpts := eval.MustBuffer(endpointConf.Remain)
+	bufferOpts := eval.MustBuffer(append(blockBodies, endpointConf.Remain)...)
 	if len(proxies)+len(requests) > 1 { // also buffer with more possible results
 		bufferOpts |= eval.BufferResponse
 	}
@@ -150,15 +155,15 @@ func newEndpointOptions(confCtx *hcl.EvalContext, endpointConf *config.Endpoint,
 	}
 
 	return &handler.EndpointOptions{
-		APIName:       apiName,
-		Context:       endpointConf.Remain,
-		Error:         errTpl,
-		LogPattern:    endpointConf.Pattern,
-		Proxies:       proxies,
-		ReqBodyLimit:  bodyLimit,
-		ReqBufferOpts: bufferOpts,
-		Requests:      requests,
-		Response:      response,
-		ServerOpts:    serverOptions,
+		APIName:      apiName,
+		Context:      endpointConf.Remain,
+		Error:        errTpl,
+		LogPattern:   endpointConf.Pattern,
+		Proxies:      proxies,
+		ReqBodyLimit: bodyLimit,
+		BufferOpts:   bufferOpts,
+		Requests:     requests,
+		Response:     response,
+		ServerOpts:   serverOptions,
 	}, nil
 }